autoscale-agent/requirements.txt
docker==7.1.0
PyYAML==6.0.2
requests==2.32.3
